refactor(backend): align quiz media modules with main layout
