+++ backend/tests/test_quiz_results.py
@@ -6,7 +6,7 @@ from fastapi.testclient import TestClient
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-from quiz_results_service import MOCK_QUESTIONS
+from services.quiz_results import MOCK_QUESTIONS
 
 
 def _full_correct_payload() -> dict:
